+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## v0.0.3 (August XX, 2020)
 
 - Better error reporting.
+- Rename project to tea-console.
 
 
 ---