+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # CHANGELOG
 
 
-## v0.0.2 (August XX, 2020)
+## v0.0.3 (August XX, 2020)
+
+- Better error reporting.
+- Rename project to tea-console.
+
+
+---
+
+
+## v0.0.2 (August 23, 2020)
 
 - Fixes.
 - Enhanced configuration.